+++ src/web/assets/src/vue/TypesenseSynonyms.vue
@@ -1,0 +1,15 @@
+<script lang="ts">
+
+// Async load the Vue 3 APIs we need from the Vue ESM
+import { defineComponent } from 'vue'
+import GridSynonymSections from "@/vue/organisms/grids/GridSynonymSections.vue";
+
+export default defineComponent({
+
+    components: {
+        'grid-synonym-sections': GridSynonymSections,
+    },
+
+});
+
+</script>